+++ flipt-client-java/src/main/java/io/flipt/client/FliptEvaluationClient.java
@@ -17,39 +17,38 @@ import java.util.Optional;
 public class FliptEvaluationClient {
   private final Pointer engine;
 
-  private final String namespace;
-
   private final ObjectMapper objectMapper;
 
   public interface CLibrary extends Library {
     CLibrary INSTANCE = (CLibrary) Native.load("fliptengine", CLibrary.class);
 
-    Pointer initialize_engine(String[] namespaces, String opts);
+    Pointer initialize_engine(String namespace, String opts);
 
-    String evaluate_boolean(Pointer engine, String evaluation_request);
+    Pointer evaluate_boolean(Pointer engine, String evaluation_request);
 
-    String evaluate_variant(Pointer engine, String evaluation_request);
+    Pointer evaluate_variant(Pointer engine, String evaluation_request);
 
-    String evaluate_batch(Pointer engine, String batch_evaluation_request);
+    Pointer evaluate_batch(Pointer engine, String batch_evaluation_request);
+
+    Pointer list_flags(Pointer engine);
 
     void destroy_engine(Pointer engine);
+
+    void destroy_string(Pointer str);
   }
 
   private FliptEvaluationClient(String namespace, EngineOpts engineOpts)
       throws JsonProcessingException {
-
-    String[] namespaces = {namespace};
 
     ObjectMapper objectMapper = new ObjectMapper();
     objectMapper.registerModule(new Jdk8Module());
 
     String engineOptions = objectMapper.writeValueAsString(engineOpts);
 
-    Pointer engine = CLibrary.INSTANCE.initialize_engine(namespaces, engineOptions);
+    Pointer engine = CLibrary.INSTANCE.initialize_engine(namespace, engineOptions);
 
     this.objectMapper = objectMapper;
     this.engine = engine;
-    this.namespace = namespace;
   }
 
   public static FliptEvaluationClientBuilder builder() {
@@ -101,8 +100,7 @@ public class FliptEvaluationClient {
     }
   }
 
-  private static class EvalRequest {
-    private final String namespaceKey;
+  private static class InternalEvaluationRequest {
 
     private final String flagKey;
 
@@ -110,17 +108,10 @@ public class FliptEvaluationClient {
 
     private final Map<String, String> context;
 
-    public EvalRequest(
-        String namespaceKey, String flagKey, String entityId, Map<String, String> context) {
-      this.namespaceKey = namespaceKey;
+    public InternalEvaluationRequest(String flagKey, String entityId, Map<String, String> context) {
       this.flagKey = flagKey;
       this.entityId = entityId;
       this.context = context;
-    }
-
-    @JsonProperty("namespace_key")
-    public String getNamespaceKey() {
-      return namespaceKey;
     }
 
     @JsonProperty("flag_key")
@@ -141,38 +132,40 @@ public class FliptEvaluationClient {
 
   public Result<VariantEvaluationResponse> evaluateVariant(
       String flagKey, String entityId, Map<String, String> context) throws JsonProcessingException {
-    EvalRequest evaluationRequest = new EvalRequest(this.namespace, flagKey, entityId, context);
+    InternalEvaluationRequest evaluationRequest =
+        new InternalEvaluationRequest(flagKey, entityId, context);
 
     String evaluationRequestSerialized = this.objectMapper.writeValueAsString(evaluationRequest);
-    String value = CLibrary.INSTANCE.evaluate_variant(this.engine, evaluationRequestSerialized);
+    Pointer value = CLibrary.INSTANCE.evaluate_variant(this.engine, evaluationRequestSerialized);
 
     TypeReference<Result<VariantEvaluationResponse>> typeRef =
         new TypeReference<Result<VariantEvaluationResponse>>() {};
 
-    return this.objectMapper.readValue(value, typeRef);
+    return this.readValue(value, typeRef);
   }
 
   public Result<BooleanEvaluationResponse> evaluateBoolean(
       String flagKey, String entityId, Map<String, String> context) throws JsonProcessingException {
-    EvalRequest evaluationRequest = new EvalRequest(this.namespace, flagKey, entityId, context);
+    InternalEvaluationRequest evaluationRequest =
+        new InternalEvaluationRequest(flagKey, entityId, context);
 
     String evaluationRequestSerialized = this.objectMapper.writeValueAsString(evaluationRequest);
-    String value = CLibrary.INSTANCE.evaluate_boolean(this.engine, evaluationRequestSerialized);
+    Pointer value = CLibrary.INSTANCE.evaluate_boolean(this.engine, evaluationRequestSerialized);
 
     TypeReference<Result<BooleanEvaluationResponse>> typeRef =
         new TypeReference<Result<BooleanEvaluationResponse>>() {};
-    return this.objectMapper.readValue(value, typeRef);
+    return this.readValue(value, typeRef);
   }
 
   public Result<BatchEvaluationResponse> evaluateBatch(EvaluationRequest[] batchEvaluationRequests)
       throws JsonProcessingException {
-    ArrayList<EvalRequest> evaluationRequests = new ArrayList<>(batchEvaluationRequests.length);
+    ArrayList<InternalEvaluationRequest> evaluationRequests =
+        new ArrayList<>(batchEvaluationRequests.length);
 
     for (int i = 0; i < batchEvaluationRequests.length; i++) {
       evaluationRequests.add(
           i,
-          new EvalRequest(
-              this.namespace,
+          new InternalEvaluationRequest(
               batchEvaluationRequests[i].getFlagKey(),
               batchEvaluationRequests[i].getEntityId(),
               batchEvaluationRequests[i].getContext()));
@@ -180,15 +173,33 @@ public class FliptEvaluationClient {
 
     String batchEvaluationRequestSerialized =
         this.objectMapper.writeValueAsString(evaluationRequests.toArray());
-    String value = CLibrary.INSTANCE.evaluate_batch(this.engine, batchEvaluationRequestSerialized);
+    Pointer value = CLibrary.INSTANCE.evaluate_batch(this.engine, batchEvaluationRequestSerialized);
 
     TypeReference<Result<BatchEvaluationResponse>> typeRef =
         new TypeReference<Result<BatchEvaluationResponse>>() {};
 
-    return this.objectMapper.readValue(value, typeRef);
+    return this.readValue(value, typeRef);
+  }
+
+  public Result<ArrayList<Flag>> listFlags() throws JsonProcessingException {
+    Pointer value = CLibrary.INSTANCE.list_flags(this.engine);
+
+    TypeReference<Result<ArrayList<Flag>>> typeRef =
+        new TypeReference<Result<ArrayList<Flag>>>() {};
+
+    return this.readValue(value, typeRef);
   }
 
   public void close() {
     CLibrary.INSTANCE.destroy_engine(this.engine);
+  }
+
+  private <T> T readValue(Pointer ptr, TypeReference<T> typeRef) throws JsonProcessingException {
+    try {
+      String value = ptr.getString(0, "UTF-8");
+      return this.objectMapper.readValue(value, typeRef);
+    } finally {
+      CLibrary.INSTANCE.destroy_string(ptr);
+    }
   }
 }