+++ flipt-client-python/flipt_client/models.py
@@ -1,4 +1,4 @@
-from pydantic import BaseModel
+from pydantic import BaseModel, RootModel
 from typing import List, Optional
 
 
@@ -80,4 +80,29 @@ class BooleanResult(BaseModel):
 class BatchResult(BaseModel):
     status: str
     result: Optional[BatchEvaluationResponse] = None
+    error_message: Optional[str] = None
+
+
+class Flag(BaseModel):
+    key: str
+    enabled: bool
+    type: str
+
+
+class FlagList(RootModel):
+    root: List[Flag]
+
+    def __iter__(self):
+        return iter(self.root)
+
+    def __getitem__(self, item):
+        return self.root[item]
+
+    def __len__(self):
+        return len(self.root)
+
+
+class ListFlagsResult(BaseModel):
+    status: str
+    result: Optional[FlagList] = None
     error_message: Optional[str] = None