+++ flipt-engine-ffi/examples/evaluation.rs
@@ -7,7 +7,7 @@ use std::collections::HashMap;
 
 fn main() {
     let evaluator = Evaluator::new_snapshot_evaluator(
-        vec!["default".into()],
+        "default".into(),
         HTTPParserBuilder::new("http://localhost:8080")
             .authentication(Authentication::with_client_token("secret".into()))
             .build(),
@@ -21,7 +21,6 @@ fn main() {
     let thread = std::thread::spawn(move || loop {
         std::thread::sleep(std::time::Duration::from_millis(5000));
         let variant = eng.variant(&EvaluationRequest {
-            namespace_key: "default".into(),
             flag_key: "flag1".into(),
             entity_id: "entity".into(),
             context: context.clone(),