+++ flipt-evaluation/src/lib.rs
@@ -24,7 +24,7 @@ where
     P: Parser + Send,
     S: Store + Send,
 {
-    namespaces: Vec<String>,
+    namespace: String,
     parser: P,
     store: S,
     mtx: Arc<RwLock<i32>>,
@@ -32,7 +32,6 @@ where
 
 #[repr(C)]
 pub struct EvaluationRequest {
-    pub namespace_key: String,
     pub flag_key: String,
     pub entity_id: String,
     pub context: HashMap<String, String>,
@@ -117,6 +116,8 @@ impl Default for ErrorEvaluationResponse {
     }
 }
 
+type ListFlagsResult = std::result::Result<Vec<flipt::Flag>, Error>;
+
 type VariantEvaluationResult<T> = std::result::Result<T, Error>;
 
 type BooleanEvaluationResult<T> = std::result::Result<T, Error>;
@@ -125,13 +126,13 @@ impl<P> Evaluator<P, Snapshot>
 where
     P: Parser + Send,
 {
-    pub fn new_snapshot_evaluator(namespaces: Vec<String>, parser: P) -> Result<Self, Error> {
-        let snap = Snapshot::build(&namespaces, &parser)?;
-        Ok(Evaluator::new(namespaces, parser, snap))
+    pub fn new_snapshot_evaluator(namespace: String, parser: P) -> Result<Self, Error> {
+        let snap = Snapshot::build(&namespace, &parser)?;
+        Ok(Evaluator::new(namespace, parser, snap))
     }
 
     pub fn replace_snapshot(&mut self) {
-        match Snapshot::build(&self.namespaces, &self.parser) {
+        match Snapshot::build(&self.namespace, &self.parser) {
             Ok(s) => {
                 self.replace_store(s);
             }
@@ -147,9 +148,9 @@ where
     P: Parser + Send,
     S: Store + Send,
 {
-    pub fn new(namespaces: Vec<String>, parser: P, store_impl: S) -> Self {
+    pub fn new(namespace: String, parser: P, store_impl: S) -> Self {
         Self {
-            namespaces,
+            namespace,
             parser,
             store: store_impl,
             mtx: Arc::new(RwLock::new(0)),
@@ -161,15 +162,26 @@ where
         self.store = store_impl;
     }
 
+    pub fn list_flags(&self) -> ListFlagsResult {
+        let _r_lock = self.mtx.read().unwrap();
+        match self.store.list_flags(&self.namespace) {
+            Some(f) => Ok(f),
+            None => Err(Error::Unknown(format!(
+                "failed to get flags for {}",
+                self.namespace,
+            ))),
+        }
+    }
+
     pub fn variant(
         &self,
         evaluation_request: &EvaluationRequest,
     ) -> VariantEvaluationResult<VariantEvaluationResponse> {
         let _r_lock = self.mtx.read().unwrap();
-        let flag = match self.store.get_flag(
-            &evaluation_request.namespace_key,
-            &evaluation_request.flag_key,
-        ) {
+        let flag = match self
+            .store
+            .get_flag(&self.namespace, &evaluation_request.flag_key)
+        {
             Some(f) => {
                 if f.r#type != common::FlagType::Variant {
                     return Err(Error::InvalidRequest(format!(
@@ -182,7 +194,7 @@ where
             None => {
                 return Err(Error::InvalidRequest(format!(
                     "failed to get flag information {}/{}",
-                    &evaluation_request.namespace_key, &evaluation_request.flag_key,
+                    &self.namespace, &evaluation_request.flag_key,
                 )));
             }
         };
@@ -195,10 +207,10 @@ where
         evaluation_request: &EvaluationRequest,
     ) -> BooleanEvaluationResult<BooleanEvaluationResponse> {
         let _r_lock = self.mtx.read().unwrap();
-        let flag = match self.store.get_flag(
-            &evaluation_request.namespace_key,
-            &evaluation_request.flag_key,
-        ) {
+        let flag = match self
+            .store
+            .get_flag(&self.namespace, &evaluation_request.flag_key)
+        {
             Some(f) => {
                 if f.r#type != common::FlagType::Boolean {
                     return Err(Error::InvalidRequest(format!(
@@ -211,7 +223,7 @@ where
             None => {
                 return Err(Error::InvalidRequest(format!(
                     "failed to get flag information {}/{}",
-                    &evaluation_request.namespace_key, &evaluation_request.flag_key,
+                    &self.namespace, &evaluation_request.flag_key,
                 )));
             }
         };
@@ -227,10 +239,7 @@ where
 
         let mut evaluation_responses: Vec<EvaluationResponse> = Vec::new();
         for request in requests {
-            let flag = match self
-                .store
-                .get_flag(&request.namespace_key, &request.flag_key)
-            {
+            let flag = match self.store.get_flag(&self.namespace, &request.flag_key) {
                 Some(f) => f,
                 None => {
                     evaluation_responses.push(EvaluationResponse {
@@ -239,7 +248,7 @@ where
                         variant_evaluation_response: None,
                         error_evaluation_response: Some(ErrorEvaluationResponse {
                             flag_key: request.flag_key,
-                            namespace_key: request.namespace_key,
+                            namespace_key: self.namespace.clone(),
                             reason: common::ErrorEvaluationReason::NotFound,
                         }),
                     });
@@ -295,15 +304,15 @@ where
             return Ok(variant_evaluation_response);
         }
 
-        let evaluation_rules = match self.store.get_evaluation_rules(
-            &evaluation_request.namespace_key,
-            &evaluation_request.flag_key,
-        ) {
+        let evaluation_rules = match self
+            .store
+            .get_evaluation_rules(&self.namespace, &evaluation_request.flag_key)
+        {
             Some(rules) => rules,
             None => {
                 return Err(Error::Unknown(format!(
                     "error getting evaluation rules for namespace {} and flag {}",
-                    evaluation_request.namespace_key.clone(),
+                    self.namespace.clone(),
                     evaluation_request.flag_key.clone()
                 )));
             }
@@ -353,13 +362,13 @@ where
 
             let distributions = match self
                 .store
-                .get_evaluation_distributions(&evaluation_request.namespace_key, &rule.id)
+                .get_evaluation_distributions(&self.namespace, &rule.id)
             {
                 Some(evaluation_distributions) => evaluation_distributions,
                 None => {
                     return Err(Error::Unknown(format!(
                         "error getting evaluation distributions for namespace {} and rule {}",
-                        evaluation_request.namespace_key.clone(),
+                        self.namespace.clone(),
                         rule.id.clone()
                     )))
                 }
@@ -436,15 +445,15 @@ where
         let now = SystemTime::now();
         let mut last_rank = 0;
 
-        let evaluation_rollouts = match self.store.get_evaluation_rollouts(
-            &evaluation_request.namespace_key,
-            &evaluation_request.flag_key,
-        ) {
+        let evaluation_rollouts = match self
+            .store
+            .get_evaluation_rollouts(&self.namespace, &evaluation_request.flag_key)
+        {
             Some(rollouts) => rollouts,
             None => {
                 return Err(Error::Unknown(format!(
                     "error getting evaluation rollouts for namespace {} and flag {}",
-                    evaluation_request.namespace_key.clone(),
+                    self.namespace.clone(),
                     evaluation_request.flag_key.clone()
                 )));
             }
@@ -1138,7 +1147,7 @@ mod tests {
             .returning(|_, _| Some(vec![]));
 
         let evaluator = &Evaluator {
-            namespaces: vec!["default".into()],
+            namespace: "default".into(),
             parser: test_parser,
             store: mock_store,
             mtx: Arc::new(RwLock::new(0)),
@@ -1147,7 +1156,6 @@ mod tests {
         let context: HashMap<String, String> = HashMap::new();
 
         let variant = evaluator.variant(&EvaluationRequest {
-            namespace_key: String::from("default"),
             flag_key: String::from("foo"),
             entity_id: String::from("user@flipt.io"),
             context,
@@ -1216,7 +1224,7 @@ mod tests {
             .returning(|_, _| Some(vec![]));
 
         let evaluator = &Evaluator {
-            namespaces: vec!["default".into()],
+            namespace: "default".into(),
             parser: test_parser,
             store: mock_store,
             mtx: Arc::new(RwLock::new(0)),
@@ -1227,7 +1235,6 @@ mod tests {
         context.insert(String::from("foo"), String::from("bar"));
 
         let variant = evaluator.variant(&EvaluationRequest {
-            namespace_key: String::from("default"),
             flag_key: String::from("foo"),
             entity_id: String::from("entity"),
             context,
@@ -1308,7 +1315,7 @@ mod tests {
             .returning(|_, _| Some(vec![]));
 
         let evaluator = &Evaluator {
-            namespaces: vec!["default".into()],
+            namespace: "default".into(),
             parser: test_parser,
             store: mock_store,
             mtx: Arc::new(RwLock::new(0)),
@@ -1320,7 +1327,6 @@ mod tests {
         context.insert(String::from("company"), String::from("flipt"));
 
         let variant = evaluator.variant(&EvaluationRequest {
-            namespace_key: String::from("default"),
             flag_key: String::from("foo"),
             entity_id: String::from("entity"),
             context,
@@ -1338,7 +1344,6 @@ mod tests {
         context.insert(String::from("bar"), String::from("boz"));
 
         let variant = evaluator.variant(&EvaluationRequest {
-            namespace_key: String::from("default"),
             flag_key: String::from("foo"),
             entity_id: String::from("entity"),
             context,
@@ -1427,7 +1432,7 @@ mod tests {
             });
 
         let evaluator = &Evaluator {
-            namespaces: vec!["default".into()],
+            namespace: "default".into(),
             parser: test_parser,
             store: mock_store,
             mtx: Arc::new(RwLock::new(0)),
@@ -1440,7 +1445,6 @@ mod tests {
         context.insert(String::from("admin"), String::from("true"));
 
         let variant = evaluator.variant(&EvaluationRequest {
-            namespace_key: String::from("default"),
             flag_key: String::from("foo"),
             entity_id: String::from("123"),
             context,
@@ -1528,7 +1532,7 @@ mod tests {
             });
 
         let evaluator = &Evaluator {
-            namespaces: vec!["default".into()],
+            namespace: "default".into(),
             parser: test_parser,
             store: mock_store,
             mtx: Arc::new(RwLock::new(0)),
@@ -1541,7 +1545,6 @@ mod tests {
         context.insert(String::from("admin"), String::from("true"));
 
         let variant = evaluator.variant(&EvaluationRequest {
-            namespace_key: String::from("default"),
             flag_key: String::from("foo"),
             entity_id: String::from("123"),
             context,
@@ -1634,7 +1637,7 @@ mod tests {
             });
 
         let evaluator = &Evaluator {
-            namespaces: vec!["default".into()],
+            namespace: "default".into(),
             parser: test_parser,
             store: mock_store,
             mtx: Arc::new(RwLock::new(0)),
@@ -1646,7 +1649,6 @@ mod tests {
         context.insert(String::from("foo"), String::from("bar"));
 
         let variant = evaluator.variant(&EvaluationRequest {
-            namespace_key: String::from("default"),
             flag_key: String::from("foo"),
             entity_id: String::from("1"),
             context: context.clone(),
@@ -1663,7 +1665,6 @@ mod tests {
         assert_eq!(v.segment_keys, vec![String::from("segment1")]);
 
         let variant = evaluator.variant(&EvaluationRequest {
-            namespace_key: String::from("default"),
             flag_key: String::from("foo"),
             entity_id: String::from("2"),
             context,
@@ -1767,7 +1768,7 @@ mod tests {
             });
 
         let evaluator = &Evaluator {
-            namespaces: vec!["default".into()],
+            namespace: "default".into(),
             parser: test_parser,
             store: mock_store,
             mtx: Arc::new(RwLock::new(0)),
@@ -1779,7 +1780,6 @@ mod tests {
         context.insert(String::from("foo"), String::from("bar"));
 
         let variant = evaluator.variant(&EvaluationRequest {
-            namespace_key: String::from("default"),
             flag_key: String::from("foo"),
             entity_id: String::from("1"),
             context: context.clone(),
@@ -1851,7 +1851,7 @@ mod tests {
             });
 
         let evaluator = &Evaluator {
-            namespaces: vec!["default".into()],
+            namespace: "default".into(),
             parser: test_parser,
             store: mock_store,
             mtx: Arc::new(RwLock::new(0)),
@@ -1860,7 +1860,6 @@ mod tests {
         let context: HashMap<String, String> = HashMap::new();
 
         let variant = evaluator.variant(&EvaluationRequest {
-            namespace_key: String::from("default"),
             flag_key: String::from("foo"),
             entity_id: String::from("10"),
             context: context.clone(),
@@ -1877,7 +1876,6 @@ mod tests {
         assert_eq!(v.segment_keys, vec![String::from("segment1")]);
 
         let variant = evaluator.variant(&EvaluationRequest {
-            namespace_key: String::from("default"),
             flag_key: String::from("foo"),
             entity_id: String::from("01"),
             context,
@@ -1948,7 +1946,7 @@ mod tests {
             .returning(|_, _| Some(vec![]));
 
         let evaluator = &Evaluator {
-            namespaces: vec!["default".into()],
+            namespace: "default".into(),
             parser: test_parser,
             store: mock_store,
             mtx: Arc::new(RwLock::new(0)),
@@ -1958,7 +1956,6 @@ mod tests {
         context.insert(String::from("bar"), String::from("baz"));
 
         let variant = evaluator.variant(&EvaluationRequest {
-            namespace_key: String::from("default"),
             flag_key: String::from("foo"),
             entity_id: String::from("entity"),
             context,
@@ -2039,7 +2036,7 @@ mod tests {
             .returning(|_, _| Some(vec![]));
 
         let evaluator = &Evaluator {
-            namespaces: vec!["default".into()],
+            namespace: "default".into(),
             parser: test_parser,
             store: mock_store,
             mtx: Arc::new(RwLock::new(0)),
@@ -2050,7 +2047,6 @@ mod tests {
         context.insert(String::from("company"), String::from("flipt"));
 
         let variant = evaluator.variant(&EvaluationRequest {
-            namespace_key: String::from("default"),
             flag_key: String::from("foo"),
             entity_id: String::from("entity"),
             context,
@@ -2068,7 +2064,6 @@ mod tests {
         context.insert(String::from("bar"), String::from("boz"));
 
         let variant = evaluator.variant(&EvaluationRequest {
-            namespace_key: String::from("default"),
             flag_key: String::from("foo"),
             entity_id: String::from("entity"),
             context,
@@ -2157,7 +2152,7 @@ mod tests {
             });
 
         let evaluator = &Evaluator {
-            namespaces: vec!["default".into()],
+            namespace: "default".into(),
             parser: test_parser,
             store: mock_store,
             mtx: Arc::new(RwLock::new(0)),
@@ -2169,7 +2164,6 @@ mod tests {
         context.insert(String::from("admin"), String::from("true"));
 
         let variant = evaluator.variant(&EvaluationRequest {
-            namespace_key: String::from("default"),
             flag_key: String::from("foo"),
             entity_id: String::from("123"),
             context,
@@ -2257,7 +2251,7 @@ mod tests {
             });
 
         let evaluator = &Evaluator {
-            namespaces: vec!["default".into()],
+            namespace: "default".into(),
             parser: test_parser,
             store: mock_store,
             mtx: Arc::new(RwLock::new(0)),
@@ -2269,7 +2263,6 @@ mod tests {
         context.insert(String::from("admin"), String::from("true"));
 
         let variant = evaluator.variant(&EvaluationRequest {
-            namespace_key: String::from("default"),
             flag_key: String::from("foo"),
             entity_id: String::from("123"),
             context,
@@ -2354,7 +2347,7 @@ mod tests {
             });
 
         let evaluator = &Evaluator {
-            namespaces: vec!["default".into()],
+            namespace: "default".into(),
             parser: test_parser,
             store: mock_store,
             mtx: Arc::new(RwLock::new(0)),
@@ -2365,7 +2358,6 @@ mod tests {
         context.insert(String::from("bar"), String::from("baz"));
 
         let variant = evaluator.variant(&EvaluationRequest {
-            namespace_key: String::from("default"),
             flag_key: String::from("foo"),
             entity_id: String::from("1"),
             context: context.clone(),
@@ -2382,7 +2374,6 @@ mod tests {
         assert_eq!(v.segment_keys, vec![String::from("segment1")]);
 
         let variant = evaluator.variant(&EvaluationRequest {
-            namespace_key: String::from("default"),
             flag_key: String::from("foo"),
             entity_id: String::from("2"),
             context,
@@ -2486,7 +2477,7 @@ mod tests {
             });
 
         let evaluator = &Evaluator {
-            namespaces: vec!["default".into()],
+            namespace: "default".into(),
             parser: test_parser,
             store: mock_store,
             mtx: Arc::new(RwLock::new(0)),
@@ -2497,7 +2488,6 @@ mod tests {
         context.insert(String::from("premium_user"), String::from("true"));
 
         let variant = evaluator.variant(&EvaluationRequest {
-            namespace_key: String::from("default"),
             flag_key: String::from("foo"),
             entity_id: String::from("1"),
             context: context.clone(),
@@ -2569,7 +2559,7 @@ mod tests {
             });
 
         let evaluator = &Evaluator {
-            namespaces: vec!["default".into()],
+            namespace: "default".into(),
             parser: test_parser,
             store: mock_store,
             mtx: Arc::new(RwLock::new(0)),
@@ -2578,7 +2568,6 @@ mod tests {
         let mut context: HashMap<String, String> = HashMap::new();
 
         let variant = evaluator.variant(&EvaluationRequest {
-            namespace_key: String::from("default"),
             flag_key: String::from("foo"),
             entity_id: String::from("10"),
             context: context.clone(),
@@ -2595,7 +2584,6 @@ mod tests {
         assert_eq!(v.segment_keys, vec![String::from("segment1")]);
 
         let variant = evaluator.variant(&EvaluationRequest {
-            namespace_key: String::from("default"),
             flag_key: String::from("foo"),
             entity_id: String::from("01"),
             context: context.clone(),
@@ -2613,7 +2601,6 @@ mod tests {
 
         context.insert(String::from("foo"), String::from("bar"));
         let variant = evaluator.variant(&EvaluationRequest {
-            namespace_key: String::from("default"),
             flag_key: String::from("foo"),
             entity_id: String::from("01"),
             context,
@@ -2691,7 +2678,7 @@ mod tests {
             });
 
         let evaluator = &Evaluator {
-            namespaces: vec!["default".into()],
+            namespace: "default".into(),
             parser: test_parser,
             store: mock_store,
             mtx: Arc::new(RwLock::new(0)),
@@ -2702,7 +2689,6 @@ mod tests {
         context.insert(String::from("bar"), String::from("baz"));
 
         let variant = evaluator.variant(&EvaluationRequest {
-            namespace_key: String::from("default"),
             flag_key: String::from("foo"),
             entity_id: String::from("1"),
             context: context.clone(),
@@ -2803,7 +2789,7 @@ mod tests {
             });
 
         let evaluator = &Evaluator {
-            namespaces: vec!["default".into()],
+            namespace: "default".into(),
             parser: test_parser,
             store: mock_store,
             mtx: Arc::new(RwLock::new(0)),
@@ -2814,7 +2800,6 @@ mod tests {
         context.insert(String::from("bar"), String::from("baz"));
 
         let variant = evaluator.variant(&EvaluationRequest {
-            namespace_key: String::from("default"),
             flag_key: String::from("foo"),
             entity_id: String::from("1"),
             context: context.clone(),