+++ flipt-engine-ffi/CHANGELOG.md
@@ -1,5 +1,26 @@
 # Changelog
 
+## [0.1.4](https://github.com/flipt-io/flipt-client-sdks/compare/flipt-engine-ffi-v0.1.3...flipt-engine-ffi-v0.1.4) (2024-03-05)
+
+
+### Bug Fixes
+
+* **engine-ffi:** prevent memory leaks ([#157](https://github.com/flipt-io/flipt-client-sdks/issues/157)) ([60b4ec9](https://github.com/flipt-io/flipt-client-sdks/commit/60b4ec9c51204c29d1eae403ea1612c8f6a5faab))
+
+## [0.1.3](https://github.com/flipt-io/flipt-client-sdks/compare/flipt-engine-ffi-v0.1.2...flipt-engine-ffi-v0.1.3) (2024-03-01)
+
+
+### Features
+
+* list flags + single namespace ([#151](https://github.com/flipt-io/flipt-client-sdks/issues/151)) ([cf09c85](https://github.com/flipt-io/flipt-client-sdks/commit/cf09c857a851c70aea97bccc9a159c4c7b5540a0))
+
+## [0.1.2](https://github.com/flipt-io/flipt-client-sdks/compare/flipt-engine-ffi-v0.1.1...flipt-engine-ffi-v0.1.2) (2024-02-24)
+
+
+### Miscellaneous Chores
+
+* release 0.1.2 ([fea6c5a](https://github.com/flipt-io/flipt-client-sdks/commit/fea6c5a894cb2f138dbf1bb3badc855f2910e7a4))
+
 ## [0.1.1](https://github.com/flipt-io/flipt-client-sdks/compare/flipt-engine-ffi-v0.1.0...flipt-engine-ffi-v0.1.1) (2024-02-06)